chore: split constants
This splits constants and moves decorators to separate files.
