+++ blocket_api/constants.py
@@ -1,0 +1,4 @@
+BASE_URL = "https://api.blocket.se"
+SITE_URL = "https://www.blocket.se"
+BYTBIL_URL = "https://api.bytbil.com"
+QASA_URL = "https://api.qasa.se/graphql"
